+++ docs/conf.py
@@ -14,7 +14,7 @@
 project = 'Django Provenance'
 copyright = '2025, German Aerospace Center (DLR)'
 author = 'German Aerospace Center (DLR)'
-release = '0.2.1'
+release = '0.2.2'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration